refactor(backoffice-v2): moved the task cell components to different files
