+++ apps/backoffice-v2/src/components/pages/Individual/components/Alert/interfaces.ts
@@ -1,0 +1,3 @@
+export interface IAlertProps {
+  value: string;
+}
