+++ apps/backoffice-v2/src/components/pages/Individual/components/CallToAction/interfaces.ts
@@ -1,0 +1,8 @@
+export interface ICallToActionProps {
+  value: string;
+  data: {
+    id: string;
+    disabled: boolean;
+    approvalStatus: 'rejected' | 'approved' | 'revision';
+  };
+}
