+++ apps/backoffice-v2/src/components/pages/Individual/components/Container/interfaces.ts
@@ -1,0 +1,6 @@
+export interface IContainerProps {
+  id: string;
+  value: Array<{
+    type: string;
+  }>;
+}
