+++ apps/backoffice-v2/src/components/pages/Individual/components/Details/Details.tsx
@@ -1,11 +1,12 @@
 import { ctw } from '../../../../../utils/ctw/ctw';
 import { EditableDetails } from 'components/pages/Individual/components/EditableDetails/EditableDetails';
 import { Separator } from 'components/atoms/Separator/separator';
-import React from 'react';
+import React, { FunctionComponent } from 'react';
 import { useParams } from '@tanstack/react-router';
 import { useEndUserWithWorkflowQuery } from '../../../../../lib/react-query/queries/useEndUserWithWorkflowQuery/useEndUserWithWorkflowQuery';
+import { IDetailsProps } from 'components/pages/Individual/components/Details/interfaces';
 
-export const Details = ({ id, value }) => {
+export const Details: FunctionComponent<IDetailsProps> = ({ id, value }) => {
   const { endUserId } = useParams();
   const { data: endUser } = useEndUserWithWorkflowQuery(endUserId);
