+++ apps/backoffice-v2/src/components/pages/Individual/components/Details/interfaces.ts
@@ -1,0 +1,15 @@
+export interface IDetailsProps {
+  id: string;
+  value: {
+    id: string;
+    title: string;
+    data: Array<{
+      title: string;
+      isEditable: boolean;
+      type: string;
+      format?: string;
+      pattern?: string;
+      value: unknown;
+    }>;
+  };
+}
