+++ apps/backoffice-v2/src/components/pages/Individual/components/EditableDetails/interfaces.ts
@@ -1,0 +1,18 @@
+export interface IEditableDetails {
+  data: Array<{
+    title: string;
+    value: unknown;
+    isEditable: boolean;
+    type: string;
+    format?: string;
+    pattern?: string;
+  }>;
+  valueId: string;
+  id: string;
+  documents: Array<{
+    id: string;
+    properties: Record<string, string>;
+  }>;
+  title: string;
+  workflowId: string;
+}
