+++ apps/backoffice-v2/src/components/pages/Individual/components/FaceComparison/interfaces.ts
@@ -1,0 +1,6 @@
+export interface IFaceComparisonProps {
+  value: {
+    faceAUrl: string;
+    faceBUrl: string;
+  };
+}
