+++ apps/backoffice-v2/src/components/pages/Individual/components/Heading/Heading.tsx
@@ -1,0 +1,6 @@
+import React, { FunctionComponent } from 'react';
+import { IHeadingProps } from 'components/pages/Individual/components/Heading/interfaces';
+
+export const Heading: FunctionComponent<IHeadingProps> = ({ value }) => (
+  <h2 className={`ml-2 mt-6 p-2 text-2xl font-bold`}>{value}</h2>
+);
