+++ apps/backoffice-v2/src/components/pages/Individual/components/Heading/interfaces.ts
@@ -1,0 +1,3 @@
+export interface IHeadingProps {
+  value: string;
+}
