+++ apps/backoffice-v2/src/components/pages/Individual/components/MultiDocuments/interfaces.ts
@@ -1,0 +1,9 @@
+export interface IMultiDocumentsProps {
+  value: {
+    data: Array<{
+      imageUrl: string;
+      title: string;
+      fileType: string;
+    }>;
+  };
+}
