+++ apps/backoffice-v2/src/components/pages/Individual/hooks/useIndividual/components.tsx
@@ -1,0 +1,17 @@
+import { Heading } from 'components/pages/Individual/components/Heading/Heading';
+import { Alert } from 'components/pages/Individual/components/Alert/Alert';
+import { Container } from 'components/pages/Individual/components/Container/Container';
+import { CallToAction } from 'components/pages/Individual/components/CallToAction/CallToAction';
+import { FaceComparison } from 'components/pages/Individual/components/FaceComparison/FaceComparison';
+import { Details } from 'components/pages/Individual/components/Details/Details';
+import { MultiDocuments } from 'components/pages/Individual/components/MultiDocuments/MultiDocuments';
+
+export const components = {
+  heading: Heading,
+  alert: Alert,
+  container: Container,
+  callToAction: CallToAction,
+  faceComparison: FaceComparison,
+  details: Details,
+  multiDocuments: MultiDocuments,
+};
